seperate the data validation (in data.py) from functions for hashes
